Sun, Aug 6, 2023, 1:15 AM +01:00

+++ readability/readability.c
@@ -117,6 +117,6 @@ int coleman_liau_formula(int letters,int words,int sentences)
     float L = (letters * 100) / words;
     float S = (sentences * 100) / words;
 
-    int grade = lround(0.0588 * L - 0.296 * S - 15.8);
+    float grade = lround(0.0588 * L - 0.296 * S - 15.8);
     return grade;
 }
